Updated windows project
Upgraded windows project to the VS 2022.

Fixed windows link tinyxml2 error.
Fixed empty argument menu_selector.
Fixed IGameHandler::getInstance() nullptr error.

+++ projects/NarutoSenki/Classes/StartMenu.cpp
@@ -570,7 +570,7 @@ void StartMenu::enterCustomMode()
 
 void StartMenu::enterTrainingMode()
 {
-	s_IGameModeHandler = s_ModeHandlers[(size_t)GameMode::Classic];
+	s_IGameModeHandler = s_ModeHandlers[GameMode::Classic];
 
 	enableCustomSelect = false;
 	enterSelectLayer();
